Change project structure

+++ Algorithmic Toolbox/Week 1/Stress/max_pairwise_product_stress.cpp
@@ -1,3 +1,4 @@
+
 #include <iostream>
 #include <vector>
 #include <cstdlib>
@@ -6,23 +7,22 @@ using std::vector;
 using std::cin;
 using std::cout;
 
-long long  MaxPairwiseProduct(const vector<int>& numbers) {
-    long long max_product = 0;
+int64_t  MaxPairwiseProduct(const vector<int>& numbers) {
+    int64_t max_product = 0;
     int n = numbers.size();
 
     for (int i = 0; i < n; ++i) {
         for (int j = i + 1; j < n; ++j) {
-            if((long long)numbers[i] * numbers[j] > max_product) {
+            if ((int64_t)numbers[i] * numbers[j] > max_product) {
                 max_product = numbers[i] * numbers[j];
             }
         }
-
     }
 
     return max_product;
 }
 
-long long MaxPairwiseProductFast(const vector<int>& numbers) {
+int64_t MaxPairwiseProductFast(const vector<int>& numbers) {
     int n = numbers.size();
 
     int max_index1 = -1;
@@ -39,7 +39,7 @@ long long MaxPairwiseProductFast(const vector<int>& numbers) {
            max_index2 = j;
         }
     }
-    return ((long long)(numbers[max_index1])) * numbers[max_index2];
+    return ((int64_t)(numbers[max_index1])) * numbers[max_index2];
 }
 
 
@@ -57,8 +57,8 @@ int main() {
         }
         cout << "\n";
 
-        long long result1 = MaxPairwiseProduct(a);
-        long long result2 = MaxPairwiseProductFast(a);
+        int64_t result1 = MaxPairwiseProduct(a);
+        int64_t result2 = MaxPairwiseProductFast(a);
 
         if (result1 != result2) {
             cout << "WRONG ANSWER: " << result1 << " " << result2 << "\n";
